refactor(acceptance): Decompose scripts/acceptance/scenarios/29-plot-detail-roundtrip.ts (warren-69fe)
