+++ scripts/acceptance/scenarios/lib/git-helpers.ts
@@ -1,0 +1,44 @@
+import { AcceptanceError } from "../../lib/assert.ts";
+
+export interface RunResult {
+	stdout: string;
+	stderr: string;
+}
+
+export async function runIn(
+	cwd: string,
+	cmd: readonly string[],
+	env: Record<string, string>,
+	label = "command",
+): Promise<RunResult> {
+	const proc = Bun.spawn({
+		cmd: [...cmd],
+		cwd,
+		env,
+		stdin: "ignore",
+		stdout: "pipe",
+		stderr: "pipe",
+	});
+	const [stdout, stderr, exitCode] = await Promise.all([
+		new Response(proc.stdout).text(),
+		new Response(proc.stderr).text(),
+		proc.exited,
+	]);
+	if ((exitCode ?? 0) !== 0) {
+		throw new AcceptanceError(
+			`${label} failed (${cmd.join(" ")} in ${cwd}): exit ${exitCode}\nstderr: ${stderr}\nstdout: ${stdout}`,
+		);
+	}
+	return { stdout, stderr };
+}
+
+export function withGitIdentity(): Record<string, string> {
+	return {
+		PATH: process.env.PATH ?? "",
+		HOME: process.env.HOME ?? "/tmp",
+		GIT_AUTHOR_NAME: "Warren Acceptance",
+		GIT_AUTHOR_EMAIL: "acceptance@warren.invalid",
+		GIT_COMMITTER_NAME: "Warren Acceptance",
+		GIT_COMMITTER_EMAIL: "acceptance@warren.invalid",
+	};
+}
