+++ scripts/file-size-budgets.json
@@ -8,7 +8,6 @@
 		"scripts/acceptance/scenarios/26-plan-run-roundtrip.ts": 563,
 		"scripts/acceptance/scenarios/27-plan-run-plot-roundtrip.ts": 856,
 		"scripts/acceptance/scenarios/28-plot-list-and-create.ts": 559,
-		"scripts/acceptance/scenarios/29-plot-detail-roundtrip.ts": 1013,
 		"scripts/acceptance/scenarios/31-plot-plan-run-synthesis.ts": 950,
 		"scripts/acceptance/scenarios/32-plot-workbench-loop.ts": 706,
 		"src/client/client.ts": 569,
